Final Code uploads
Hey everyone,

This is all the code you will need for the expo. There are 3 important ones, "Email.Script", "Scraper.Script" and "Scraper_1". 

"Email.Script" is the Code that I created to send out E-mails. 

"Scraper.Script" is my final attempt at getting the webscraper to work on the ryanair site. Unfortunatly I couldn't manage to get it done.

"Scraper_1" is my proud success of a running webscraper. It scraping information of a different website and isn't perfect, but it works.

Best of luck tomorrow 

cheers, Greg